backend/requirements.txt
APScheduler==3.11.0
Flask==3.1.1
flask_cors==6.0.0
flask_limiter==3.12
flask_pymongo==3.0.1
google_api_python_client==2.172.0
humanize==4.12.3
isodate==0.7.2
itsdangerous==2.2.0
nltk==3.9.1
numpy>=1.21
redis>=3.0
openai==1.93.1
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
Requests==2.32.4
resend==2.10.0
Werkzeug==3.1.3
youtube_transcript_api==1.1.1